requirements.txt
botbuilder-integration-aiohttp>=4.14.0
psycopg2
fastapi
uvicorn
psycopg2
requests